requirements.txt
Django~=3.1
django-jet
six
djangorestframework